Changed all money-type from int to BigDecimal. Fixed the circular reference by @Lazy for Bean.

+++ src/main/java/org/sorokin_school/AccountProperties.java
@@ -3,12 +3,14 @@ package org.sorokin_school;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class AccountProperties {
     @Value("${account.default-amount:500}")
-    private int defaultAmount;
+    private BigDecimal defaultAmount;
 
-    public int getDefaultAmount() {
+    public BigDecimal getDefaultAmount() {
         return defaultAmount;
     }
