+++ src/main/java/org/sorokin_school/model/Account.java
@@ -1,11 +1,13 @@
 package org.sorokin_school.model;
 
+import java.math.BigDecimal;
+
 public class Account {
     private int id;
     private int userId;
-    private int moneyAmount;
+    private BigDecimal moneyAmount;
 
-    public Account(int id, int userId, int moneyAmount) {
+    public Account(int id, int userId, BigDecimal moneyAmount) {
         this.id = id;
         this.userId = userId;
         this.moneyAmount = moneyAmount;
@@ -19,7 +21,7 @@ public class Account {
         return userId;
     }
 
-    public int getMoneyAmount() {
+    public BigDecimal getMoneyAmount() {
         return moneyAmount;
     }
 
@@ -31,7 +33,7 @@ public class Account {
         this.userId = userId;
     }
 
-    public void setMoneyAmount(int moneyAmount) {
+    public void setMoneyAmount(BigDecimal moneyAmount) {
         this.moneyAmount = moneyAmount;
     }
